Move Webpay::Base -> WebpayPlus::Base
- Previously the Base class was under Transbank::Webpay, but that makes
no sense since Webpay is a product family instead of a product.
- The base class has been moved to its corresponding product,
WebpayPlus: Transbank::Webpay::WebpayPlus::Base
